pyproject.toml
[tool.poetry]
name = "user-app"
version = "0.1.0"
description = ""
authors = ["feba66 <git@feba66.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyargon2 = "^1.0.7"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max-line-length = 160